+++ scripts/macros/spells/dragonsBreath.js
@@ -15,6 +15,12 @@ export async function dragonsBreath({speaker, actor, token, character, item, arg
     ];
     featureData.system.save.dc = chris.getSpellDC(this.item);
     featureData.system.description.value = chris.getItemDescription('CPR - Descriptions', 'Dragon Breath');
+    featureData.flags['chris-premades'] = {
+		'spell': {
+			'castData': this.castData
+		}
+	}
+	featureData.flags['chris-premades'].spell.castData.school = this.item.system.school;
     async function effectMacro () {
 		await warpgate.revert(token.document, 'Dragon Breath');
 	}